TASK: Replace TYPO3.Party model with custom implementation
This replaces the usage of the `TYPO3.Party` package in favor of
a new `User` domain model.
This serves as a basis for further adjustments, namely the synchronization
of attributes such that users can update their profiles.

This also comes with some minor tweaks that are not exactly related but
too small to be pushed as separate PRs.